chore: rename to PaymentMethodHandlerService
