+++ packages/backend/src/index.ts
@@ -48,7 +48,7 @@ import { createAutoPeeringService } from './payment-method/ilp/auto-peering/serv
 import { createAutoPeeringRoutes } from './payment-method/ilp/auto-peering/routes'
 import axios from 'axios'
 import { createIlpPaymentService } from './payment-method/ilp/service'
-import { createPaymentMethodManagerService } from './payment-method/service'
+import { createPaymentMethodHandlerService } from './payment-method/handler/service'
 
 BigInt.prototype.toJSON = function () {
   return this.toString()
@@ -431,8 +431,8 @@ export function initIocContainer(
     })
   })
 
-  container.singleton('paymentMethodManagerService', async (deps) => {
-    return createPaymentMethodManagerService({
+  container.singleton('paymentMethodHandlerService', async (deps) => {
+    return createPaymentMethodHandlerService({
       logger: await deps.use('logger'),
       knex: await deps.use('knex'),
       ilpPaymentService: await deps.use('ilpPaymentService')
@@ -447,7 +447,7 @@ export function initIocContainer(
       receiverService: await deps.use('receiverService'),
       feeService: await deps.use('feeService'),
       paymentPointerService: await deps.use('paymentPointerService'),
-      paymentMethodManagerService: await deps.use('paymentMethodManagerService')
+      paymentMethodHandlerService: await deps.use('paymentMethodHandlerService')
     })
   })
